+++ src/classes/parsing/ui5class/ts/CustomTSObject.ts
@@ -29,6 +29,7 @@ export class CustomTSObject extends AbstractCustomClass<
 	ObjectLiteralElementLike,
 	ObjectLiteralExpression
 > {
+	defaultModelClassName?: string;
 	parentClassNameDotNotation = "";
 	readonly typeChecker: TypeChecker;
 	protected _fillIsAbstract(): void {